Insert all functions of helper in userauth_test
